fix(rbac): keep the git-derived gates out of the integration lanes
The legacy-cohort gates read git history to derive the initial commit's object
vocabulary and grants. They sat in the backend root package with no build
constraint, so `go test ./...` ran them in EVERY lane — including the twelve
integration shards and the integration job, which clone shallow. All of them
went red with the gate's own "this needs full history" message.

The design said this plainly: nothing on the integration side may depend on git
history, which is why the fixture is committed and the derivation stays in the
unit lane. The intent was right and the implementation did not carry it — a
package with no tag is not a lane.

Constrain the two git-reading files to `!integration`, the exact mirror of the
tag the integration lanes build with. The unit lane runs them; the integration
lanes no longer compile them, so there is nothing to fail. The policy.go
vocabulary parser moves to its own untagged file, because rbacvocabulary_test.go
reads the working tree rather than history and legitimately runs in both lanes.

Teach the shard discovery about the new constraint. It refuses a build tag it
cannot statically decide rather than guessing — correctly, since a tag it
mis-decided would leave tests in no slice while its ran==assigned teeth saw
nothing missing. `!integration` is decidable for the same reason `integration`
is: this lane's build always satisfies it.

The fetch-depth additions from the previous commit stay. extension-reference and
integration-unit-coverage both run the UNIT lane (`make test` / `go test ./...`,
no tags), so they still compile these gates and still need the history.
