+++ backend/rbaclegacydocument_test.go
@@ -1,6 +1,8 @@
 // SPDX-License-Identifier: BUSL-1.1
 // SPDX-FileCopyrightText: 2026 Gradion
 
+//go:build !integration
+
 package backendarch
 
 // Deriving the legacy cohort's NAMES from history (rbaclegacyinstall_test.go)
