+++ scripts/test-integration-parallel.sh
@@ -153,6 +153,13 @@ if (( SHARD_TOTAL > 0 )); then
         [[ -n "$expr" ]] || continue
         case "$expr" in
           integration) ;;
+          # !integration is the mirror of `integration`: the unit-lane-only
+          # gates. The RBAC legacy-cohort gates read git history to derive the
+          # initial commit's vocabulary, and the integration lanes clone
+          # shallow, so those files are excluded from THIS build exactly as the
+          # compiler excludes them. Statically decidable for the same reason
+          # `integration` is — this lane's build always satisfies the tag.
+          '!integration') skip_file=1 ;;
           e2e_llm|livesmoke|voicelive) skip_file=1 ;;
           *)
             echo "FAIL: $f carries build constraint '$expr' — not one the static shard discovery knows" >&2
